# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Edgebook",
    "description": "College Football Paper-Betting Platform API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/accounts": {
      "post": {
        "tags": [
          "accounts"
        ],
        "summary": "Create Account Endpoint",
        "description": "Create an active fictional account and optionally credit its opening bankroll.",
        "operationId": "create_account_endpoint_accounts_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AccountCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AccountResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/accounts/{account_id}/bets": {
      "post": {
        "tags": [
          "simulated-wagers"
        ],
        "summary": "Place Bet Endpoint",
        "operationId": "place_bet_endpoint_accounts__account_id__bets_post",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "idempotency-key",
            "in": "header",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "maxLength": 100
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BetCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetPlacementResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "simulated-wagers"
        ],
        "summary": "List Bets Endpoint",
        "operationId": "list_bets_endpoint_accounts__account_id__bets_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          },
          {
            "name": "market_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Market Type"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetPage"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_id}/bets/{bet_id}": {
      "get": {
        "tags": [
          "simulated-wagers"
        ],
        "summary": "Get Bet Endpoint",
        "operationId": "get_bet_endpoint_accounts__account_id__bets__bet_id__get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "bet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BetResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/accounts/{account_id}/bets/{bet_id}/review": {
      "get": {
        "tags": [
          "simulated-wagers"
        ],
        "summary": "Get Review Endpoint",
        "operationId": "get_review_endpoint_accounts__account_id__bets__bet_id__review_get",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "bet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bet Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "simulated-wagers"
        ],
        "summary": "Complete Review Endpoint",
        "description": "Complete a local human rationale review.",
        "operationId": "complete_review_endpoint_accounts__account_id__bets__bet_id__review_put",
        "parameters": [
          {
            "name": "account_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Account Id"
            }
          },
          {
            "name": "bet_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Bet Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewComplete"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReviewResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AccountCreate": {
        "properties": {
          "owner_name": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Owner Name"
          },
          "starting_bankroll": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Starting Bankroll",
            "default": "0.00"
          }
        },
        "type": "object",
        "required": [
          "owner_name"
        ],
        "title": "AccountCreate",
        "description": "Payload for creating a fictional simulation-credit account."
      },
      "AccountResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "owner_name": {
            "type": "string",
            "title": "Owner Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "starting_bankroll": {
            "type": "string",
            "title": "Starting Bankroll"
          },
          "current_balance": {
            "type": "string",
            "title": "Current Balance"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "owner_name",
          "is_active",
          "starting_bankroll",
          "current_balance",
          "created_at",
          "updated_at"
        ],
        "title": "AccountResponse",
        "description": "Public representation of a fictional account."
      },
...
      "BetCreate": {
        "properties": {
          "market_id": {
            "type": "string",
            "title": "Market Id"
          },
          "selection": {
            "$ref": "#/components/schemas/MarketSelection"
          },
          "quote_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Id"
          },
          "stake": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "string"
              }
            ],
            "title": "Stake"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rationale_category": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/RationaleCategory"
              },
              {
                "type": "null"
              }
            ]
          },
          "notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          }
        },
        "type": "object",
        "required": [
          "market_id",
          "selection",
          "stake"
        ],
        "title": "BetCreate"
      },
      "BetPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/BetResponse"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "BetPage"
      },
      "BetPlacementResponse": {
        "properties": {
          "bet": {
            "$ref": "#/components/schemas/BetResponse"
          },
          "current_balance": {
            "type": "string",
            "title": "Current Balance"
          }
        },
        "type": "object",
        "required": [
          "bet",
          "current_balance"
        ],
        "title": "BetPlacementResponse"
      },
      "BetResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "game_id": {
            "type": "string",
            "title": "Game Id"
          },
          "market_id": {
            "type": "string",
            "title": "Market Id"
          },
          "quote_id": {
            "type": "string",
            "title": "Quote Id"
          },
          "stake_transaction_id": {
            "type": "string",
            "title": "Stake Transaction Id"
          },
          "payout_transaction_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payout Transaction Id"
          },
          "selection": {
            "$ref": "#/components/schemas/MarketSelection"
          },
          "market_type": {
            "$ref": "#/components/schemas/MarketType"
          },
          "line": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "american_odds": {
            "type": "integer",
            "title": "American Odds"
          },
          "quote_source": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Source"
          },
          "quote_source_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Source Id"
          },
          "quote_observed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Quote Observed At"
          },
          "stake": {
            "type": "string",
            "title": "Stake"
          },
          "bankroll_before": {
            "type": "string",
            "title": "Bankroll Before"
          },
          "payout": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Payout"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "rationale_category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Rationale Category"
          },
          "notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Notes"
          },
          "status": {
            "$ref": "#/components/schemas/BetStatus"
          },
          "placed_at": {
            "type": "string",
            "title": "Placed At"
          },
          "settled_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Settled At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "account_id",
          "game_id",
          "market_id",
          "quote_id",
          "stake_transaction_id",
          "payout_transaction_id",
          "selection",
          "market_type",
          "line",
          "american_odds",
          "quote_source",
          "quote_source_id",
          "quote_observed_at",
          "stake",
          "bankroll_before",
          "payout",
          "reason",
          "rationale_category",
          "notes",
          "status",
          "placed_at",
          "settled_at"
        ],
        "title": "BetResponse"
      },
      "BetStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "WON",
          "LOST",
          "PUSH"
        ],
        "title": "BetStatus"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "MarketSelection": {
        "type": "string",
        "enum": [
          "HOME",
          "AWAY",
          "OVER",
          "UNDER"
        ],
        "title": "MarketSelection",
        "description": "The available outcome selections across the supported market types."
      },
...
      "MarketType": {
        "type": "string",
        "enum": [
          "SPREAD",
          "MONEYLINE",
          "TOTAL"
        ],
        "title": "MarketType",
        "description": "Supported college-football market categories."
      },
...
      "RationaleCategory": {
        "type": "string",
        "enum": [
          "MATCHUP_ANALYSIS",
          "STATISTICAL_EDGE",
          "LINE_VALUE",
          "INJURY_IMPACT",
          "SITUATIONAL",
          "CONTRARIAN",
          "OTHER"
        ],
        "title": "RationaleCategory",
        "description": "Typed categories for bet rationale, enabling analytics by reasoning type."
      },
...
      "ReviewComplete": {
        "properties": {
          "reviewer_label": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Reviewer Label"
          },
          "summary": {
            "type": "string",
            "maxLength": 2000,
            "minLength": 1,
            "title": "Summary"
          },
          "bias_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "maxItems": 20,
            "title": "Bias Flags"
          },
          "assessment_notes": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Notes"
          },
          "review_version": {
            "type": "string",
            "maxLength": 64,
            "minLength": 1,
            "title": "Review Version",
            "default": "human-v1"
          }
        },
        "type": "object",
        "required": [
          "reviewer_label",
          "summary"
        ],
        "title": "ReviewComplete"
      },
...
      "ReviewResponse": {
        "properties": {
          "bet_id": {
            "type": "string",
            "title": "Bet Id"
          },
          "status": {
            "$ref": "#/components/schemas/ReviewStatus"
          },
          "reviewer_label": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reviewer Label"
          },
          "summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Summary"
          },
          "bias_flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Bias Flags"
          },
          "assessment_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assessment Notes"
          },
          "review_version": {
            "type": "string",
            "title": "Review Version"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "started_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Started At"
          },
          "completed_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          }
        },
        "type": "object",
        "required": [
          "bet_id",
          "status",
          "reviewer_label",
          "summary",
          "bias_flags",
          "assessment_notes",
          "review_version",
          "created_at",
          "started_at",
          "completed_at"
        ],
        "title": "ReviewResponse"
      },
      "ReviewStatus": {
        "type": "string",
        "enum": [
          "PENDING",
          "IN_REVIEW",
          "COMPLETED",
          "FAILED",
          "NOT_APPLICABLE"
        ],
        "title": "ReviewStatus",
        "description": "Lifecycle for asynchronous human rationale review."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}